Fix B-Tree off by one errors
Tackle one problem at a time. The tree is implemented in-memory for now.
Once the remaining bugs are fixed, we'll move to an on-disk model as per
the project spec. The only problem left that I see is the distinct word
count is incorrect, we seem to be double counting some words?